Confirmation on delete, toast messages and notification bell icon done

+++ resources/views/assemblies/index.blade.php
@@ -22,11 +22,6 @@
                                     Assembly</a>
                             </div>
                         @endcan
-                        @if ($message = Session::get('success'))
-                            <div class="alert alert-success">
-                                <p>{{ $message }}</p>
-                            </div>
-                        @endif
                         <div class="cus_msg_div">
                         </div>
                         <div class="card-body px-4 pb-2">
@@ -78,7 +73,15 @@
     </main>
     <x-plugins></x-plugins>
 </x-layout>
-
+@if ($message = Session::get('success'))
+    <script>
+        var message = "{{ $message }}";
+        jQuery('#toast_body_msg').html(message); 
+        let myAlert = document.querySelector('.toast');
+        let bsAlert = new  bootstrap.Toast(myAlert);
+        bsAlert.show();
+    </script>
+@endif
 <script type="text/javascript">
     var $ = jQuery.noConflict();
     var permission_delete = "{{ checkPermission('assembly-delete') }}";
@@ -143,6 +146,7 @@
                     orderable: false,
                     searchable: false,
                     render: function(data, type, full, meta) {
+                        var confirmation = "'Are you sure you want to delete?'";
                         var btn =
                             '<a rel="tooltip" class="btn btn-info btn-link m-2" href="assemblies/show/' +
                             full.id +
@@ -155,7 +159,7 @@
                         }
                         if (permission_delete == "granted") {
                             btn +=
-                                '<a rel="tooltip" class="btn btn-danger btn-link m-2" href="assemblies/destroy/' +
+                                '<a rel="tooltip" onclick="return confirm('+confirmation+')" class="btn btn-danger btn-link m-2" href="assemblies/destroy/' +
                                 full.id +
                                 '" data-original-title="Delete Assembly" title="Delete Assembly"><i class="material-icons">delete</i><div class="ripple-container"></div></a>';
                         }
