+++ resources/views/events/index.blade.php
@@ -21,11 +21,6 @@
                                         class="material-icons text-sm">add</i>&nbsp;&nbsp;Add New Event</a>
                             </div>
                         @endcan
-                        @if ($message = Session::get('success'))
-                            <div class="alert alert-success">
-                                <p>{{ $message }}</p>
-                            </div>
-                        @endif
                         <div class="cus_msg_div">
                         </div>
                         <div class="card-body px-4 pb-2">
@@ -62,8 +57,16 @@
         </div>
     </main>
     <x-plugins></x-plugins>
-
 </x-layout>
+@if ($message = Session::get('success'))
+    <script>
+        var message = "{{ $message }}";
+        jQuery('#toast_body_msg').html(message); 
+        let myAlert = document.querySelector('.toast');
+        let bsAlert = new  bootstrap.Toast(myAlert);
+        bsAlert.show();
+    </script>
+@endif
 <script type="text/javascript">
     var $ = jQuery.noConflict();
     var permission_delete = "{{ checkPermission('election_info-delete') }}";
@@ -109,6 +112,7 @@
                     orderable: false,
                     searchable: false,
                     render: function(data, type, full, meta) {
+                        var confirmation = "'Are you sure you want to delete?'";
                         var btn =
                             '<a rel="tooltip" class="btn btn-info btn-link m-2" href="event/show/' +
                             full.id +
@@ -121,7 +125,7 @@
                         }
                         if (permission_delete == "granted") {
                             btn +=
-                                '<a rel="tooltip" class="btn btn-danger btn-link m-2" href="event/destroy/' +
+                                '<a rel="tooltip" onclick="return confirm('+confirmation+')" class="btn btn-danger btn-link m-2" href="event/destroy/' +
                                 full.id +
                                 '" data-original-title="Delete Event" title="Delete Event"><i class="material-icons">delete</i><div class="ripple-container"></div></a>';
                         }
