+++ resources/views/users/index.blade.php
@@ -22,11 +22,6 @@
                                 User</a>
                         </div>
                         @endcan
-                        @if ($message = Session::get('success'))
-                            <div class="alert alert-success">
-                                <p>{{ $message }}</p>
-                            </div>
-                        @endif
                         <div class="cus_msg_div">
                         </div>
                         <div class="card-body px-4 pb-2">
@@ -74,6 +69,15 @@
     </main>
     <x-plugins></x-plugins>
 </x-layout>
+@if ($message = Session::get('success'))
+    <script>
+        var message = "{{ $message }}";
+        jQuery('#toast_body_msg').html(message); 
+        let myAlert = document.querySelector('.toast');
+        let bsAlert = new  bootstrap.Toast(myAlert);
+        bsAlert.show();
+    </script>
+@endif
 <script>
 var $ = jQuery.noConflict();
 var permission_delete = "{{ checkPermission('user-delete') }}";
@@ -129,12 +133,13 @@ var permission_edit = "{{ checkPermission('user-edit') }}";
                         orderable: false,
                         searchable: false,
                         render: function(data, type, full, meta) {
+                            var confirmation = "'Are you sure you want to delete?'";
                             var btn = '<a rel="tooltip" class="btn btn-info btn-link m-2" href="users/'+full.id+'" data-original-title="Show User" title="Show User"><i class="material-icons">visibility</i><div class="ripple-container"></div></a>';
                             if(permission_edit == "granted"){
                                 btn += '<a rel="tooltip" class="btn btn-success btn-link m-2" href="users/'+full.id+'/edit" data-original-title="Edit User" title="Edit User"><i class="material-icons">edit</i><div class="ripple-container"></div></a>';
                             }
                             if(permission_delete == "granted"){
-                                btn += '<a rel="tooltip" class="btn btn-danger btn-link m-2" href="user/delete/'+full.id+'" data-original-title="Delete User" title="Delete User"><i class="material-icons">delete</i><div class="ripple-container"></div></a>';
+                                btn += '<a rel="tooltip" onclick="return confirm('+confirmation+')" class="btn btn-danger btn-link m-2" href="user/delete/'+full.id+'" data-original-title="Delete User" title="Delete User"><i class="material-icons">delete</i><div class="ripple-container"></div></a>';
                             }
                             return btn;
                         }
